--- app/src/main/java/com/sun/moviedb/view/detail/DetailMovieActivity.kt
@@ -1,19 +1,0 @@
-package com.sun.moviedb.view.detail
-
-import android.os.Bundle
-import com.sun.moviedb.R
-import com.sun.moviedb.base.BaseActivity
-
-/**
- * Created by nguyenxuanhoi on 2019-08-23.
- * @author nguyen.xuan.hoi@sun-asterisk.com
- */
-class DetailMovieActivity : BaseActivity() {
-    override val getContentViewId = R.layout.activity_detail
-
-    override fun initializeView(savedInstanceState: Bundle?) {
-    }
-
-    override fun initializeComponents() {
-    }
-}
