+++ app/src/main/java/com/sun/moviedb/view/home/HomeViewModel.kt
@@ -18,19 +18,17 @@ import com.sun.moviedb.utils.CategoryQuery
 class HomeViewModel(
         application: Application,
         private val homeRepository: HomeRepository,
-        private val reponsitory: MovieByCategoryRepository)
+        private val repository: MovieByCategoryRepository)
     : AndroidViewModel(application) {
 
-    private val _genres = homeRepository.getAllGenres()
+    private val _genres by lazy {
+        homeRepository.getAllGenres()
+    }
     val genres: LiveData<BaseResponse<List<Genres>>>
         get() = _genres
 
-    companion object {
-        private val FIRST_PAGE = 1
-    }
-
     private val _categories by lazy {
-        reponsitory.getQueryTypeCategory()
+        repository.getQueryTypeCategory()
     }
 
     fun getCategories(): List<CategoryDTO> {
@@ -38,31 +36,22 @@ class HomeViewModel(
     }
 
     private val _moviesPopular: LiveData<BaseResponse<List<Movie>>> by lazy {
-        reponsitory.getMoviesByCategory(CategoryQuery.POPULAR, FIRST_PAGE)
+        repository.getMoviesByCategory(CategoryQuery.POPULAR, FIRST_PAGE)
     }
     val moviesPopular: LiveData<BaseResponse<List<Movie>>>
         get() = _moviesPopular
 
     private val _moviesUpComing: LiveData<BaseResponse<List<Movie>>> by lazy {
-        reponsitory.getMoviesByCategory(CategoryQuery.UP_COMING, FIRST_PAGE)
+        repository.getMoviesByCategory(CategoryQuery.UP_COMING, FIRST_PAGE)
     }
     val moviesUpComming: LiveData<BaseResponse<List<Movie>>>
         get() = _moviesUpComing
 
-    private val _moviesNowPlaying: LiveData<BaseResponse<List<Movie>>> by lazy {
-        reponsitory.getMoviesByCategory(CategoryQuery.NOW_PLAYING, FIRST_PAGE)
-    }
-    val moviesNowPlaying: LiveData<BaseResponse<List<Movie>>>
-        get() = _moviesNowPlaying
-
-    private val _moviesTopRate: LiveData<BaseResponse<List<Movie>>> by lazy {
-        reponsitory.getMoviesByCategory(CategoryQuery.TOP_RATE, FIRST_PAGE)
-    }
-    val movieTopRate: LiveData<BaseResponse<List<Movie>>>
-        get() = _moviesTopRate
-
     override fun onCleared() {
         super.onCleared()
         homeRepository.coroutines.onClear()
+    }
+    companion object {
+        private val FIRST_PAGE = 1
     }
 }
