+++ app/src/main/res/layout/fragment_home.xml
@@ -43,13 +43,16 @@
                 android:background="@drawable/bg_layout"
                 app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-                <fragment
-                    android:id="@+id/hostHomeNav"
-                    android:name="androidx.navigation.fragment.NavHostFragment"
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/recyclerView"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    app:defaultNavHost="true"
-                    app:navGraph="@navigation/nv_home" />
+                    app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent"
+                    tools:listitem="@layout/item_home_movie" />
             </LinearLayout>
 
             <com.google.android.material.appbar.AppBarLayout
