+++ app/src/main/res/layout/fragment_movie_by_category.xml
@@ -11,5 +11,5 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:listitem="@layout/item_home_movie" />
+        tools:listitem="@layout/item_movie" />
 </layout>
