--- app/src/main/res/navigation/nv_detail.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nv_detail"
-    app:startDestination="@id/detailMovieFragment">
-
-    <fragment
-        android:id="@+id/detailMovieFragment"
-        android:name="com.sun.moviedb.view.detail.DetailMovieFragment"
-        android:label="DetailMovieFragment" />
-</navigation>
