--- app/src/main/res/navigation/nv_home.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nv_home"
-    app:startDestination="@id/discoverFragment">
-
-    <fragment
-        android:id="@+id/discoverFragment"
-        android:name="com.sun.moviedb.view.home.discover.DiscoverFragment"
-        android:label="DiscoverFragment">
-        <action
-            android:id="@+id/actDiscoverToMovieByCategory"
-            app:destination="@id/movieByCategoryFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right" />
-        <action
-            android:id="@+id/actDiscoverToDetailActivity"
-            app:destination="@id/detailMovieActivity" />
-
-
-    </fragment>
-    <fragment
-        android:id="@+id/movieByCategoryFragment"
-        android:name="com.sun.moviedb.view.category.MovieByCategoryFragment"
-        android:label="MovieByCategoryFragment">
-        <argument
-            android:name="query"
-            app:argType="string" />
-        <action
-            android:id="@+id/actMovieByCategoryToDetailActivity"
-            app:destination="@id/detailMovieActivity"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-    <fragment
-        android:id="@+id/movieByGenresFragment"
-        android:name="com.sun.moviedb.view.genres.MovieByGenresFragment"
-        android:label="MovieByGenresFragment">
-
-        <action
-            android:id="@+id/actMovieByGenresToDetailActivity"
-            app:destination="@id/detailMovieActivity"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-    <fragment
-        android:id="@+id/searchFragment"
-        android:name="com.sun.moviedb.view.search.SearchFragment"
-        android:label="SearchFragment">
-        <action
-            android:id="@+id/actSearchToDetailActivity"
-            app:destination="@id/detailMovieActivity"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-    <activity
-        android:id="@+id/detailMovieActivity"
-        android:name="com.sun.moviedb.view.detail.DetailMovieActivity"
-        android:label="DetailMovieActivity" />
-
-</navigation>
