+++ app/src/main/res/values-night/colors.xml
@@ -4,6 +4,5 @@
     <color name="color_primary_dark">#95DBFF</color>
     <color name="color_accent">@android:color/black</color>
     <color name="color_on_primary">@android:color/black</color>
-    <color name="color_background">@android:color/black</color>
     <color name="color_status_bar">@android:color/black</color>
 </resources>
