+++ app/src/main/res/values/colors.xml
@@ -4,7 +4,6 @@
     <color name="color_primary_dark">#95DBFF</color>
     <color name="color_accent">#6ACAFD</color>
     <color name="color_on_primary">@android:color/white</color>
-    <color name="color_background">@android:color/white</color>
     <color name="color_status_bar">#95DBFF</color>
-
+    <color name="color_yellow">#FF9800</color>
 </resources>
