+++ app/src/main/res/values/dimens.xml
@@ -16,5 +16,10 @@
     <dimen name="dp_120">120dp</dimen>
     <dimen name="dp_170">170dp</dimen>
     <dimen name="dp_12">12dp</dimen>
+    <dimen name="dp_10">10dp</dimen>
+    <dimen name="dp_130">130dp</dimen>
+    <dimen name="dp_160">160dp</dimen>
+    <dimen name="sp_16">16sp</dimen>
+    <dimen name="dp_110">110dp</dimen>
 
 </resources>
